move type declarations from types.go to indiviual files
