+++ flow/input.go
@@ -1,5 +1,15 @@
 package flow // import "github.com/orkestr8/xgraph/flow"
 
+import (
+	xg "github.com/orkestr8/xgraph"
+)
+
+type input struct {
+	edges   xg.EdgeSlice
+	recv    []<-chan work
+	collect chan work
+}
+
 func (input *input) run() {
 	for _, c := range input.recv {
 		go func(cc <-chan work) {
